API codegen: Change behavior of a map index with a missing key.
Previously indexing a map with a non-existent key would return a
“zero-initialized” value that was _not added_ to the map. The
new implementation is to add the zero-initialized value to the map
and return the reference to that.

This has the consequence that maps can now grow when indexing.

Justification:

This is a step towards allowing the user to write expressions like:

  m[k].field = 1

where the map value is a non-ref struct type.

C++ and the new jit-executor both allow this form, but the go codegen
does not as map indexed values are immutable and cannot be referenced.

Longer-term we probably want to throw a error (abort) if the key does
not already exist. Implementing this is trivial for C++ and the new
jit-executor, but the gles.api requires work to support this change
(non-existent keys are indexed).

Change-Id: I56fc59f88644f2567909f4cdc2e7644d4d21e63b